Support a new config tag, muchsync.and_tags
Muchsync ORs most tags when syncing, but ANDs tags in in new.tags.
Initial-tagging systems like [afew][] use the new.tags configuration
element in a different way: they have notmuch assign a tag "new", and
then run afew in a post-notmuch hook.  The post-notmuch hook assigns the
"inbox" tag.  Muchsync should therefore treat this as a new tag, and AND
it when syncing, even though it's not in the new.tags configuration
element.

[afew]: https://github.com/teythoon/afew